feat(files.service): add extra-security on delete
In principle, this is not necessary and will probably never be used, but it ensures even more so that no important folder can be accidentally deleted.

+++ src/controller.ts
@@ -425,6 +425,11 @@ export class Controller {
   }
 
   private deleteFolder(folder: IFolder): void {
+    if (!this.fileService.isSafeToDelete(folder.path)) {
+      this.printError('Folder no safe to delete');
+      return;
+    }
+
     this.fileService
       .deleteDir(folder.path)
       .then(response => {
